refactor: change components names in proper way
